// src/types/SoundType.ts
export enum SoundType {

  SE_01 = '<se.1>',
  SE_02 = '<se.2>',
  SE_03 = '<se.3>',
  SE_04 = '<se.4>',
  SE_05 = '<se.5>',
  SE_06 = '<se.6>',
  SE_07 = '<se.7>',
  SE_08 = '<se.8>',
  SE_09 = '<se.9>',
  SE_10 = '<se.10>',
  SE_11 = '<se.11>',
  SE_12 = '<se.12>',
  SE_13 = '<se.13>',
  SE_14 = '<se.14>',
  SE_15 = '<se.15>',
  SE_16 = '<se.16>',

  CONFIRM = 'confirm',
  CANCEL = 'cancel',
  ENTER_CHAT = 'enter_chat',

  SILENT = 'silent',

}

export const soundTypes = [
  SoundType.SE_01, SoundType.SE_02, SoundType.SE_03, SoundType.SE_04,
  SoundType.SE_05, SoundType.SE_06, SoundType.SE_07, SoundType.SE_08,
  SoundType.SE_09, SoundType.SE_10, SoundType.SE_11, SoundType.SE_12,
  SoundType.SE_13, SoundType.SE_14, SoundType.SE_15, SoundType.SE_16,
];